+++ flutter/example/lib/main.dart
@@ -21,36 +21,39 @@ const String _exampleDsn =
 final _channel = const MethodChannel('example.flutter.sentry.io');
 
 Future<void> main() async {
-  await SentryFlutter.init(
-    (options) {
-      options.dsn = _exampleDsn;
-      options.tracesSampleRate = 1.0;
-      options.reportPackages = false;
-      options.addInAppInclude('sentry_flutter_example');
-      options.considerInAppFramesByDefault = false;
-      options.attachThreads = true;
-      options.enableWindowMetricBreadcrumbs = true;
-      options.addIntegration(LoggingIntegration());
-      options.sendDefaultPii = true;
-      options.reportSilentFlutterErrors = true;
-      options.enableNdkScopeSync = true;
-
-      // options.attachScreenshot = true;
-      // We can enable Sentry debug logging during development. This is likely
-      // going to log too much for your app, but can be useful when figuring out
-      // configuration issues, e.g. finding out why your events are not uploaded.
-      options.debug = true;
-    },
-    // Init your App.
-    appRunner: () => runApp(
-      SentryUserInteractionWidget(
-        child: DefaultAssetBundle(
-          bundle: SentryAssetBundle(enableStructuredDataTracing: true),
-          child: MyApp(),
+  await setupSentry(() => runApp(
+        // SentryScreenshotWidget(
+        SentryUserInteractionWidget(
+          child: DefaultAssetBundle(
+            bundle: SentryAssetBundle(enableStructuredDataTracing: true),
+            child: MyApp(),
+          ),
         ),
-      ),
-    ),
-  );
+      ));
+}
+
+Future<void> setupSentry(AppRunner appRunner) async {
+  await SentryFlutter.init((options) {
+    options.dsn = _exampleDsn;
+    options.tracesSampleRate = 1.0;
+    options.reportPackages = false;
+    options.addInAppInclude('sentry_flutter_example');
+    options.considerInAppFramesByDefault = false;
+    options.attachThreads = true;
+    options.enableWindowMetricBreadcrumbs = true;
+    options.addIntegration(LoggingIntegration());
+    options.sendDefaultPii = true;
+    options.reportSilentFlutterErrors = true;
+    options.enableNdkScopeSync = true;
+    // options.attachScreenshot = true;
+
+    // We can enable Sentry debug logging during development. This is likely
+    // going to log too much for your app, but can be useful when figuring out
+    // configuration issues, e.g. finding out why your events are not uploaded.
+    options.debug = true;
+  },
+      // Init your App.
+      appRunner: appRunner);
 }
 
 class MyApp extends StatefulWidget {